# services/api_gateway/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IM 即时通讯系统 API",
    "description": "基于微服务架构的即时通讯系统 REST API 文档\n\n## 认证说明\n除了 `/api/auth/*` 开头的接口，其他所有接口都需要在 Header 中携带 JWT Token：\n```\nAuthorization: Bearer <your_access_token>\n```\n\n## 错误码说明\n- `401`: 未授权，Token 无效或已过期\n- `400`: 请求参数错误\n- `500`: 服务器内部错误",
    "version": "1.0.0",
    "contact": {
      "name": "IM Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "用户注册",
        "description": "创建新用户账号",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 32,
                    "example": "testuser"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "123456"
                  },
                  "display_name": {
                    "type": "string",
                    "example": "测试用户"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或用户名已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "用户登录",
        "description": "使用用户名和密码登录，获取 JWT Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "testuser"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "刷新 Token",
        "description": "使用 Refresh Token 获取新的 Access Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refresh_token"
                ],
                "properties": {
                  "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取当前用户资料",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserProfile"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "更新当前用户资料",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_name": {
                    "type": "string",
                    "example": "新昵称"
                  },
                  "avatar_url": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserProfile"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/contacts": {
      "get": {
        "tags": [
          "联系人"
        ],
        "summary": "获取联系人列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserBrief"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/contacts/apply": {
      "post": {
        "tags": [
          "联系人"
        ],
        "summary": "发送好友申请",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "target_user_id"
                ],
                "properties": {
                  "target_user_id": {
                    "type": "integer",
                    "example": 2
                  },
                  "remark": {
                    "type": "string",
                    "example": "你好"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "申请发送成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/contacts/handle": {
      "post": {
        "tags": [
          "联系人"
        ],
        "summary": "处理好友申请",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "target_user_id"
                ],
                "properties": {
                  "target_user_id": {
                    "type": "integer",
                    "example": 2
                  },
                  "accept": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "处理成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/conversations": {
      "get": {
        "tags": [
          "会话"
        ],
        "summary": "获取会话列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Conversation"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "post": {
        "tags": [
          "会话"
        ],
        "summary": "创建会话",
        "description": "创建单聊或群聊会话",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "member_ids"
                ],
                "properties": {
                  "type": {
                    "type": "integer",
                    "example": 1,
                    "description": "1: 单聊, 2: 群聊"
                  },
                  "title": {
                    "type": "string",
                    "example": "项目群"
                  },
                  "member_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              },
              "examples": {
                "单聊": {
                  "value": {
                    "type": 1,
                    "member_ids": [
                      2
                    ]
                  }
                },
                "群聊": {
                  "value": {
                    "type": 2,
                    "member_ids": [
                      2,
                      3,
                      4
                    ],
                    "title": "测试群"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "$ref": "#/components/schemas/Conversation"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
...
    "/api/contacts/{id}": {
      "delete": {
        "tags": [
          "联系人"
        ],
        "summary": "删除联系人",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "目标用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "输入从 /api/auth/login 获取的 access_token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "error message"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refresh_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "expires_in": {
            "type": "integer",
            "example": 900,
            "description": "Access Token 过期时间（秒）"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfile"
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "status": {
            "type": "string",
            "example": "active"
          }
        }
      }
    }
  }
}